backend/ocr/easyocr_app.py
"""
EasyOCR App for Qualcomm AI Hub Models
Provides the EasyOCRApp class needed for OCR processing with Qualcomm models
"""

import numpy as np
from PIL import Image
import cv2
from typing import List, Tuple, Any

class EasyOCRApp:
    """EasyOCR application class for Qualcomm AI Hub models."""
    
    def __init__(self, detector_session, recognizer_session, lang_list=['en']):
        self.detector_session = detector_session
        self.recognizer_session = recognizer_session
        self.lang_list = lang_list
    
    def preprocess_image(self, image_path: str) -> np.ndarray:
        """Preprocess image for OCR detection."""
        image = cv2.imread(image_path)
        if image is None:
            raise ValueError(f"Could not load image: {image_path}")
        return image
    
    def predict_text_from_image(self, image_path: str) -> List[str]:
        """Predict text from image using Qualcomm models."""
        try:
            # Load and preprocess image
            image = self.preprocess_image(image_path)
            
            # For now, return a placeholder result
            # TODO: Implement actual detection and recognition with Qualcomm models
            return ["Text detected from image"]
            
        except Exception as e:
            print(f"Error in predict_text_from_image: {e}")
            return [f"OCR Error: {str(e)}"]

class EasyOCRApp_ort:
    """EasyOCR application class for ONNX Runtime with Qualcomm models."""
    
    def __init__(self, detector_session, recognizer_session):
        self.detector_session = detector_session
        self.recognizer_session = recognizer_session
    
    def detector_preprocess(self, image_path: str) -> np.ndarray:
        """Preprocess image for detector."""
        image = cv2.imread(image_path)
        if image is None:
            raise ValueError(f"Could not load image: {image_path}")
        
        # Basic preprocessing - resize and normalize
        image = cv2.resize(image, (640, 640))
        image = image.astype(np.float32) / 255.0
        image = np.transpose(image, (2, 0, 1))  # HWC to CHW
        image = np.expand_dims(image, axis=0)  # Add batch dimension
        return image
    
    def detector_inference(self, processed_image: np.ndarray) -> List[np.ndarray]:
        """Run detector inference."""
        try:
            outputs = self.detector_session.run(None, {"input": processed_image})
            return outputs
        except Exception as e:
            print(f"Detector inference error: {e}")
            return []
    
    def detector_postprocess(self, outputs: List[np.ndarray], original_image: Image.Image) -> List[List[int]]:
        """Postprocess detector outputs to get bounding boxes."""
        # Placeholder implementation
        # TODO: Implement actual postprocessing for Qualcomm detector
        return [[100, 100, 200, 200]]  # Example bounding box
    
    def crop_and_prepare_region(self, image: Image.Image, box: List[int]) -> np.ndarray:
        """Crop and prepare region for recognition."""
        try:
            # Crop the region
            x1, y1, x2, y2 = box
            cropped = image.crop((x1, y1, x2, y2))
            
            # Convert to numpy array and preprocess
            cropped_array = np.array(cropped)
            cropped_array = cv2.resize(cropped_array, (100, 32))  # Resize for recognizer
            cropped_array = cropped_array.astype(np.float32) / 255.0
            cropped_array = np.transpose(cropped_array, (2, 0, 1))  # HWC to CHW
            cropped_array = np.expand_dims(cropped_array, axis=0)  # Add batch dimension
            return cropped_array
        except Exception as e:
            print(f"Error in crop_and_prepare_region: {e}")
            return np.zeros((1, 3, 32, 100), dtype=np.float32)
    
    def recognizer_postprocess(self, outputs: List[np.ndarray], char_list: List[str]) -> str:
        """Postprocess recognizer outputs to get text."""
        try:
            # Placeholder implementation
            # TODO: Implement actual postprocessing for Qualcomm recognizer
            return "Detected Text"
        except Exception as e:
            print(f"Recognizer postprocess error: {e}")
            return "Error"
    
    def draw_boxes_on_image(self, image: Image.Image, boxes: List[List[int]], output_path: str):
        """Draw bounding boxes on image."""
        try:
            # Convert PIL image to OpenCV format
            img_array = np.array(image)
            img_array = cv2.cvtColor(img_array, cv2.COLOR_RGB2BGR)
            
            # Draw boxes
            for box in boxes:
                x1, y1, x2, y2 = box
                cv2.rectangle(img_array, (x1, y1), (x2, y2), (0, 255, 0), 2)
            
            # Save image
            cv2.imwrite(output_path, img_array)
        except Exception as e:
            print(f"Error drawing boxes: {e}") 